move usecase to separate file
